This is what I have after experimenting all night
This is not submittable. I'm not pushing this to master because it only works with heavily borrowed code/ideas. You can play with it if you want by running the Client class, there are a few different ways to set up a game in the Game class. It is set up as player vs hard ai currently, with easy mode, and ai vs ai commented out. The alpha beta ai on hard is pretty good, fair warning. Writing all the code for move generation and position tracking etc just proved to be more than I could pull off in a night, so I just tried to get something running that implements as much as possible from our design documents. While I ended up with something that runs and was a great learning experience to put together, I regret not producing anything ready for submission. Have a great break everyone